+++ CHANGELOG.md
@@ -7,6 +7,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.6.1] - 2019-11-23
+
+### Removed
+
+- Dependency on dotenv package
+
+### Changed
+
+- Generated dates by `new` command to `YYYY-MM-DD` format.
+
 ## [0.6.0] - 2019-07-16
 
 ### Added